+++ pytket/binders/circuit/boxes.cpp
@@ -487,6 +487,29 @@ void init_boxes(py::module &m) {
           "get_impl_diag", &MultiplexedU2Box::get_impl_diag,
           ":return: flag indicating whether to implement the final diagonal "
           "gate.");
+  py::class_<
+      MultiplexedTensoredU2Box, std::shared_ptr<MultiplexedTensoredU2Box>, Op>(
+      m, "MultiplexedTensoredU2Box",
+      "A user-defined multiplexed tensor product of U2 gates specified by a "
+      "map from bitstrings to lists of :py:class:`Op`s")
+      .def(
+          py::init<const ctrl_tensored_op_map_t &>(),
+          "Construct from a map from bitstrings to equal-sized lists of "
+          ":py:class:`Op`s. "
+          "Only supports single qubit unitary gate types and "
+          ":py:class:`Unitary1qBox`.\n\n"
+          ":param op_map: Map from bitstrings to lists of :py:class:`Op`s",
+          py::arg("op_map"))
+      .def(
+          "get_circuit",
+          [](MultiplexedTensoredU2Box &box) { return *box.to_circuit(); },
+          ":return: the :py:class:`Circuit` described by the box")
+      .def(
+          "get_op_map",
+          [](MultiplexedTensoredU2Box &box) {
+            return cast_keys_to_tuples(box.get_op_map());
+          },
+          ":return: the underlying op map");
   py::class_<StatePreparationBox, std::shared_ptr<StatePreparationBox>, Op>(
       m, "StatePreparationBox",
       "A box for preparing quantum states using multiplexed-Ry and "