+++ pytket/conanfile.py
@@ -32,10 +32,10 @@ class pytketRecipe(ConanFile):
         cmake.install()
 
     def requirements(self):
-        self.requires("tket/1.2.3@tket/stable")
+        self.requires("tket/1.2.5@tket/stable")
         self.requires("tklog/0.3.3@tket/stable")
         self.requires("tkassert/0.3.3@tket/stable")
-        self.requires("symengine/0.9.0")
+        self.requires("symengine/0.10.1")
         self.requires("gmp/6.2.1")
         self.requires("pybind11/2.10.3")
         self.requires("nlohmann_json/3.11.2")