+++ tket/test/src/Circuit/test_Multiplexor.cpp
@@ -34,14 +34,14 @@ static bool check_multiplexor(
   // Also assumes op_map is not empty
   auto first_op = op_map.begin();
   unsigned n_ctrl_bits = (unsigned)first_op->first.size();
-  unsigned total_ops = 1 << n_ctrl_bits;
+  unsigned long long total_ops = 1ULL << n_ctrl_bits;
   unsigned n_targets = first_op->second->n_qubits();
   std::vector<unsigned> target_qubits(n_targets);
   std::iota(std::begin(target_qubits), std::end(target_qubits), 0);
-  unsigned block_size = 1 << n_targets;
+  unsigned long long block_size = 1ULL << n_targets;
   Eigen::MatrixXcd correct_u = Eigen::MatrixXcd::Identity(
       total_ops * block_size, total_ops * block_size);
-  for (unsigned i = 0; i < total_ops; i++) {
+  for (unsigned long long i = 0; i < total_ops; i++) {
     // Find the binary rep for i
     std::vector<bool> bin = dec_to_bin(i, n_ctrl_bits);
     auto it = op_map.find(bin);
@@ -50,6 +50,40 @@ static bool check_multiplexor(
       Op_ptr op = it->second;
       Circuit c(n_targets);
       c.add_op<unsigned>(op, target_qubits);
+      c.decompose_boxes_recursively();
+      Eigen::MatrixXcd block_m = tket_sim::get_unitary(c);
+      correct_u.block(i * block_size, i * block_size, block_size, block_size) =
+          block_m;
+    }
+  }
+  Circuit circ_copy(circ);
+  circ_copy.decompose_boxes_recursively();
+  Eigen::MatrixXcd circ_u = tket_sim::get_unitary(circ_copy);
+  return (correct_u - circ_u).cwiseAbs().sum() < ERR_EPS;
+}
+
+static bool check_multiplexor(
+    const ctrl_tensored_op_map_t &op_map, const Circuit &circ) {
+  // Assume op_map is valid
+  auto first_op = op_map.begin();
+  unsigned n_ctrl_bits = (unsigned)first_op->first.size();
+  unsigned n_targets = (unsigned)first_op->second.size();
+  unsigned long long total_ops = 1ULL << n_ctrl_bits;
+  std::vector<unsigned> target_qubits(n_targets);
+  std::iota(std::begin(target_qubits), std::end(target_qubits), 0);
+  unsigned long long block_size = 1ULL << n_targets;
+  Eigen::MatrixXcd correct_u = Eigen::MatrixXcd::Identity(
+      total_ops * block_size, total_ops * block_size);
+  for (unsigned long long i = 0; i < total_ops; i++) {
+    // Find the binary rep for i
+    std::vector<bool> bin = dec_to_bin(i, n_ctrl_bits);
+    auto it = op_map.find(bin);
+    if (it != op_map.end()) {
+      // get the matrix for op
+      Circuit c(n_targets);
+      for (unsigned j = 0; j < n_targets; j++) {
+        c.add_op<unsigned>(it->second[j], {j});
+      }
       c.decompose_boxes_recursively();
       Eigen::MatrixXcd block_m = tket_sim::get_unitary(c);
       correct_u.block(i * block_size, i * block_size, block_size, block_size) =
@@ -214,7 +248,7 @@ SCENARIO("MultiplexedU2Box decomposition", "[boxes]") {
       REQUIRE(
           (cmd.get_op_ptr()->get_type() == OpType::Unitary1qBox ||
            cmd.get_op_ptr()->get_type() == OpType::CX ||
-           cmd.get_op_ptr()->get_type() == OpType::MultiplexedRotationBox));
+           cmd.get_op_ptr()->get_type() == OpType::DiagonalBox));
     }
     REQUIRE(check_multiplexor(op_map, *c));
   }
@@ -231,12 +265,12 @@ SCENARIO("MultiplexedU2Box decomposition", "[boxes]") {
     MultiplexedU2Box multiplexor(op_map);
     std::shared_ptr<Circuit> c = multiplexor.to_circuit();
     std::vector<Command> cmds = c->get_commands();
-    REQUIRE(cmds.size() == 7 + 2);
+    REQUIRE(cmds.size() == 7 + 1);
     for (auto cmd : cmds) {
       REQUIRE(
           (cmd.get_op_ptr()->get_type() == OpType::Unitary1qBox ||
            cmd.get_op_ptr()->get_type() == OpType::CX ||
-           cmd.get_op_ptr()->get_type() == OpType::MultiplexedRotationBox));
+           cmd.get_op_ptr()->get_type() == OpType::DiagonalBox));
     }
     REQUIRE(check_multiplexor(op_map, *c));
   }
@@ -252,12 +286,31 @@ SCENARIO("MultiplexedU2Box decomposition", "[boxes]") {
     MultiplexedU2Box multiplexor(op_map);
     std::shared_ptr<Circuit> c = multiplexor.to_circuit();
     std::vector<Command> cmds = c->get_commands();
-    REQUIRE(cmds.size() == 63 + 5);
+    REQUIRE(cmds.size() == 63 + 1);
     for (auto cmd : cmds) {
       REQUIRE(
           (cmd.get_op_ptr()->get_type() == OpType::Unitary1qBox ||
            cmd.get_op_ptr()->get_type() == OpType::CX ||
-           cmd.get_op_ptr()->get_type() == OpType::MultiplexedRotationBox));
+           cmd.get_op_ptr()->get_type() == OpType::DiagonalBox));
+    }
+    REQUIRE(check_multiplexor(op_map, *c));
+  }
+  GIVEN("Random MultiplexedU2Box") {
+    ctrl_op_map_t op_map;
+    for (unsigned i = 0; i < (1 << 5); i++) {
+      Unitary1qBox m(random_unitary(2, i));
+      Op_ptr mbox_op = std::make_shared<Unitary1qBox>(m);
+      op_map.insert({dec_to_bin(i, 5), mbox_op});
+    }
+    MultiplexedU2Box multiplexor(op_map);
+    std::shared_ptr<Circuit> c = multiplexor.to_circuit();
+    std::vector<Command> cmds = c->get_commands();
+    REQUIRE(cmds.size() == 63 + 1);
+    for (auto cmd : cmds) {
+      REQUIRE(
+          (cmd.get_op_ptr()->get_type() == OpType::Unitary1qBox ||
+           cmd.get_op_ptr()->get_type() == OpType::CX ||
+           cmd.get_op_ptr()->get_type() == OpType::DiagonalBox));
     }
     REQUIRE(check_multiplexor(op_map, *c));
   }
@@ -276,7 +329,8 @@ SCENARIO("Exception handling", "[boxes]") {
     ctrl_op_map_t op_map;
     REQUIRE_THROWS_MATCHES(
         MultiplexorBox(op_map), std::invalid_argument,
-        MessageContains("No Ops provided"));
+        MessageContains(
+            "The op_map argument passed to MultiplexorBox cannot be empty."));
   }
   GIVEN("Classical wire") {
     Circuit c0(2, 1);
@@ -289,49 +343,52 @@ SCENARIO("Exception handling", "[boxes]") {
         {{1, 0}, get_op_ptr(OpType::TK2, std::vector<Expr>{0.2, 0.4, 0.4})}};
     REQUIRE_THROWS_MATCHES(
         MultiplexorBox(op_map), BadOpType,
-        MessageContains("Quantum control of classical wires not supported"));
+        MessageContains("Multiplexed operations cannot have classical wires"));
   }
   GIVEN("Bitstrings are too long") {
     std::vector<bool> bits(33);
     ctrl_op_map_t op_map = {{bits, get_op_ptr(OpType::Rx, 1.4)}};
     REQUIRE_THROWS_MATCHES(
         MultiplexedRotationBox(op_map), std::invalid_argument,
-        MessageContains("Bitstrings longer than 32 are not supported"));
+        MessageContains("supports bitstrings up to 32 bits"));
   }
   GIVEN("Unmatched bitstrings") {
     ctrl_op_map_t op_map = {
         {{0, 1}, get_op_ptr(OpType::H)}, {{1}, get_op_ptr(OpType::X)}};
     REQUIRE_THROWS_MATCHES(
         MultiplexorBox(op_map), std::invalid_argument,
-        MessageContains("Bitstrings must have the same width"));
+        MessageContains(
+            "bitstrings passed to the multiplexor must have the same width"));
   }
   GIVEN("Unmatched op sizes") {
     ctrl_op_map_t op_map = {
         {{0, 1}, get_op_ptr(OpType::H)}, {{1, 0}, get_op_ptr(OpType::CX)}};
     REQUIRE_THROWS_MATCHES(
         MultiplexorBox(op_map), std::invalid_argument,
-        MessageContains("Ops must have the same width"));
+        MessageContains("Multiplexed operations must have the same width"));
   }
   GIVEN("Mixed rotation axis") {
     ctrl_op_map_t op_map = {
         {{1}, get_op_ptr(OpType::Rz, 0.3)}, {{0}, get_op_ptr(OpType::Rx, 1.4)}};
     REQUIRE_THROWS_MATCHES(
         MultiplexedRotationBox(op_map), std::invalid_argument,
-        MessageContains("Ops must have the same rotation type"));
+        MessageContains("Ops passed to MultiplexedRotationBox must have the "
+                        "same rotation type"));
   }
   GIVEN("Non-rotation type") {
     ctrl_op_map_t op_map = {{{1}, get_op_ptr(OpType::H)}};
     REQUIRE_THROWS_MATCHES(
         MultiplexedRotationBox(op_map), BadOpType,
-        MessageContains("Ops must be either Rx, Ry, or Rz"));
+        MessageContains("Ops passed to MultiplexedRotationBox must be either "
+                        "Rx, Ry, or Rz"));
   }
   GIVEN("MultiplexedU2Box unsupported gate") {
     ctrl_op_map_t op_map = {
         {{0, 1}, get_op_ptr(OpType::H)}, {{1, 0}, get_op_ptr(OpType::CX)}};
     REQUIRE_THROWS_MATCHES(
         MultiplexedU2Box(op_map), BadOpType,
-        MessageContains(
-            "Ops must be single-qubit unitary gate types or Unitary1qBox"));
+        MessageContains("Ops passed to MultiplexedU2Box must be single-qubit "
+                        "unitary gate types or Unitary1qBox"));
   }
   GIVEN("Decompose symbolic MultiplexedU2Box") {
     Sym a = SymTable::fresh_symbol("a");
@@ -413,5 +470,163 @@ SCENARIO("MultiplexorBox Dagger & transpose", "[boxes]") {
   REQUIRE(check_multiplexor(op_map, d->transpose()));
 }
 
+SCENARIO("Simple MultiplexedTensoredU2Box decomposition", "[boxes]") {
+  ctrl_tensored_op_map_t op_map;
+  op_map.insert({{0, 0}, {get_op_ptr(OpType::X), get_op_ptr(OpType::X)}});
+  MultiplexedTensoredU2Box multiplexor(op_map);
+  std::shared_ptr<Circuit> c = multiplexor.to_circuit();
+  REQUIRE(check_multiplexor(op_map, *c));
+}
+
+SCENARIO("Random MultiplexedTensoredU2Box decomposition", "[boxes]") {
+  unsigned n_controls = 2;
+  unsigned n_targets = 2;
+  GIVEN("Random (1,1) multiplexor") {
+    n_controls = 1;
+    n_targets = 1;
+  }
+  GIVEN("Random (1,2) multiplexor") {
+    n_controls = 1;
+    n_targets = 2;
+  }
+  GIVEN("Random (2,2) multiplexor") {
+    n_controls = 2;
+    n_targets = 2;
+  }
+  GIVEN("Random (2,3) multiplexor") {
+    n_controls = 2;
+    n_targets = 3;
+  }
+  GIVEN("Random (3,2) multiplexor") {
+    n_controls = 3;
+    n_targets = 2;
+  }
+  GIVEN("Random (4,4) multiplexor") {
+    n_controls = 4;
+    n_targets = 4;
+  }
+  ctrl_tensored_op_map_t op_map;
+  unsigned seed = 0;
+  for (unsigned long long i = 0; i < (1ULL << n_controls); i++) {
+    std::vector<Op_ptr> ops;
+    for (unsigned j = 0; j < n_targets; j++) {
+      Unitary1qBox m(random_unitary(2, seed++));
+      Op_ptr op = std::make_shared<Unitary1qBox>(m);
+      ops.push_back(op);
+    }
+    op_map.insert({dec_to_bin(i, n_controls), ops});
+  }
+  MultiplexedTensoredU2Box multiplexor(op_map);
+  std::shared_ptr<Circuit> c = multiplexor.to_circuit();
+  std::vector<Command> cmds = c->get_commands();
+  REQUIRE(
+      cmds.size() ==
+      ((1ULL << (n_controls + 1)) - 1) * n_targets + n_targets + 1);
+  for (auto cmd : cmds) {
+    REQUIRE(
+        (cmd.get_op_ptr()->get_type() == OpType::Unitary1qBox ||
+         cmd.get_op_ptr()->get_type() == OpType::CX ||
+         cmd.get_op_ptr()->get_type() == OpType::MultiplexedRotationBox ||
+         cmd.get_op_ptr()->get_type() == OpType::DiagonalBox));
+  }
+  REQUIRE(check_multiplexor(op_map, *c));
+}
+SCENARIO("Test MultiplexedTensoredU2Box utilities", "[boxes]") {
+  GIVEN("symbol_substitution") {
+    Sym a = SymTable::fresh_symbol("a");
+    Expr expr_a(a);
+    ctrl_tensored_op_map_t op_map = {
+        {{0},
+         {get_op_ptr(OpType::Rz, expr_a), get_op_ptr(OpType::Rx, expr_a)}}};
+    ctrl_tensored_op_map_t num_op_map = {
+        {{0}, {get_op_ptr(OpType::Rz, 1.34), get_op_ptr(OpType::Rx, 1.34)}}};
+    MultiplexedTensoredU2Box multiplexor(op_map);
+    SymEngine::map_basic_basic smap;
+    smap[a] = Expr(1.34);
+    const MultiplexedTensoredU2Box new_box =
+        static_cast<const MultiplexedTensoredU2Box &>(
+            *multiplexor.symbol_substitution(smap));
+
+    std::shared_ptr<Circuit> c = new_box.to_circuit();
+    REQUIRE(check_multiplexor(num_op_map, *c));
+  }
+  GIVEN("free_symbols") {
+    Sym a = SymTable::fresh_symbol("a");
+    Sym b = SymTable::fresh_symbol("b");
+    Expr expr_a(a);
+    Expr expr_b(b);
+    ctrl_tensored_op_map_t op_map = {
+        {{0, 1},
+         {get_op_ptr(OpType::Rz, expr_a), get_op_ptr(OpType::Rx, expr_a)}},
+        {{1, 1}, {get_op_ptr(OpType::Rz, expr_b), get_op_ptr(OpType::X)}},
+        {{1, 0}, {get_op_ptr(OpType::Rz, expr_a), get_op_ptr(OpType::X)}}};
+    MultiplexedTensoredU2Box multiplexor(op_map);
+    const SymSet symbols = multiplexor.free_symbols();
+    REQUIRE(symbols.size() == 2);
+    REQUIRE(symbols.find(a) != symbols.end());
+    REQUIRE(symbols.find(b) != symbols.end());
+  }
+  GIVEN("Rotation Dagger & transpose") {
+    ctrl_tensored_op_map_t op_map = {
+        {{0, 1}, {get_op_ptr(OpType::Rz, 3.7), get_op_ptr(OpType::X)}},
+        {{1, 1}, {get_op_ptr(OpType::Rz, 1), get_op_ptr(OpType::H)}},
+        {{1, 0}, {get_op_ptr(OpType::Rz, 2.5), get_op_ptr(OpType::H)}}};
+    MultiplexedTensoredU2Box multiplexor(op_map);
+    // Test dagger
+    const MultiplexedTensoredU2Box dag_box =
+        static_cast<const MultiplexedTensoredU2Box &>(*multiplexor.dagger());
+    std::shared_ptr<Circuit> c = dag_box.to_circuit();
+    REQUIRE(check_multiplexor(op_map, c->dagger()));
+    // Test transpose
+    const MultiplexedTensoredU2Box transpose_box =
+        static_cast<const MultiplexedTensoredU2Box &>(*multiplexor.transpose());
+    std::shared_ptr<Circuit> d = transpose_box.to_circuit();
+    REQUIRE(check_multiplexor(op_map, d->transpose()));
+  }
+}
+SCENARIO("Test MultiplexedTensoredU2Box exceptions", "[boxes]") {
+  GIVEN("Empty op_map") {
+    ctrl_tensored_op_map_t op_map;
+    REQUIRE_THROWS_MATCHES(
+        MultiplexedTensoredU2Box(op_map), std::invalid_argument,
+        MessageContains(
+            "The op_map argument passed to MultiplexedTensoredU2Box cannot be "
+            "empty."));
+  }
+  GIVEN("Bitstrings are too long") {
+    std::vector<bool> bits(33);
+    ctrl_tensored_op_map_t op_map = {{bits, {get_op_ptr(OpType::Rx, 1.4)}}};
+    REQUIRE_THROWS_MATCHES(
+        MultiplexedTensoredU2Box(op_map), std::invalid_argument,
+        MessageContains(
+            "MultiplexedTensoredU2Box only supports bitstrings up to 32 bits"));
+  }
+  GIVEN("Unmatched bitstrings") {
+    ctrl_tensored_op_map_t op_map = {
+        {{0, 1}, {get_op_ptr(OpType::H)}}, {{1}, {get_op_ptr(OpType::X)}}};
+    REQUIRE_THROWS_MATCHES(
+        MultiplexedTensoredU2Box(op_map), std::invalid_argument,
+        MessageContains("The bitstrings passed to MultiplexedTensoredU2Box "
+                        "must have the same width."));
+  }
+  GIVEN("Unmatched op sizes") {
+    ctrl_tensored_op_map_t op_map = {
+        {{0, 1}, {get_op_ptr(OpType::H)}},
+        {{1, 0}, {get_op_ptr(OpType::X), get_op_ptr(OpType::X)}}};
+    REQUIRE_THROWS_MATCHES(
+        MultiplexedTensoredU2Box(op_map), std::invalid_argument,
+        MessageContains(
+            "Each tensored operation passed to MultiplexedTensoredU2Box must "
+            "have the same number of U2 components"));
+  }
+  GIVEN("Unsupported gate") {
+    ctrl_tensored_op_map_t op_map = {
+        {{0, 1}, {get_op_ptr(OpType::H)}}, {{1, 0}, {get_op_ptr(OpType::CX)}}};
+    REQUIRE_THROWS_MATCHES(
+        MultiplexedTensoredU2Box(op_map), BadOpType,
+        MessageContains("Ops passed to MultiplexedTensoredU2Box must be "
+                        "single-qubit unitary gate types or Unitary1qBox"));
+  }
+}
 }  // namespace test_Multiplexor
 }  // namespace tket